+++ src/super/i-block/CHANGELOG.md
@@ -9,6 +9,12 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v3.0.0-rc.??? (2021-??-??)
+
+#### :house: Internal
+
+* Added API based on the `v-hook` directive to attach hook listeners with functional and flyweight components
+
 ## v3.0.0-rc.121 (2021-01-12)
 
 #### :memo: Documentation